doc:安全检测菜单下的页面常量修改引用位置
