+++ app/renderer/src/main/src/defaultConstants/NewPortScan.ts
@@ -1,0 +1,5 @@
+import {ScanPortPageInfoProps} from "@/store/pageInfo"
+
+export const defaultScanPortPageInfo: ScanPortPageInfoProps = {
+    targets: ""
+}
