+++ app/renderer/src/main/src/defaultConstants/SimpleDetect.ts
@@ -1,0 +1,3 @@
+import {SimpleDetectPageInfoProps} from "@/store/pageInfo"
+
+export const defaultSimpleDetectPageInfo: SimpleDetectPageInfoProps = {}
