+++ app/renderer/src/main/src/defaultConstants/SpaceEnginePage.ts
@@ -1,0 +1,3 @@
+import {SpaceEnginePageInfoProps} from "@/store/pageInfo"
+
+export const defaultSpaceEnginePageInfo: SpaceEnginePageInfoProps = {}
