+++ app/renderer/src/main/src/defaultConstants/YakPoC.ts
@@ -1,0 +1,10 @@
+import {PocPageInfoProps} from "@/store/pageInfo"
+
+export const defaultPocPageInfo: PocPageInfoProps = {
+    selectGroup: [],
+    selectGroupListByKeyWord: [],
+    formValue: {},
+    https: false,
+    httpFlowIds: [],
+    request: new Uint8Array()
+}
