+++ app/renderer/src/main/src/pages/layout/mainOperatorContent/MainOperatorContent.tsx
@@ -76,18 +76,7 @@ import emiter from "@/utils/eventBus/eventBus"
 import {shallow} from "zustand/shallow"
 import {menuBodyHeight} from "@/pages/globalVariable"
 import {RemoteGV} from "@/yakitGV"
-import {
-    PageNodeItemProps,
-    PageProps,
-    defPage,
-    defaultBrutePageInfo,
-    defaultPocPageInfo,
-    defaultScanPortPageInfo,
-    saveFuzzerCache,
-    defaultSimpleDetectPageInfo,
-    defaultSpaceEnginePageInfo,
-    usePageInfo
-} from "@/store/pageInfo"
+import {PageNodeItemProps, PageProps, defPage, saveFuzzerCache, usePageInfo} from "@/store/pageInfo"
 import {startupDuplexConn, closeDuplexConn} from "@/utils/duplex/duplex"
 import cloneDeep from "lodash/cloneDeep"
 import {onToManageGroup} from "@/pages/securityTool/yakPoC/YakPoC"
@@ -95,7 +84,15 @@ import {apiFetchQueryYakScriptGroupLocal} from "@/pages/plugins/utils"
 import {PluginGroupType} from "@/pages/plugins/group/PluginGroups"
 import {ExpandAndRetractExcessiveState} from "@/pages/plugins/operator/expandAndRetract/ExpandAndRetract"
 import {DefFuzzerTableMaxData, defaultAdvancedConfigValue, defaultPostTemplate} from "@/defaultConstants/HTTPFuzzerPage"
-import {defPluginBatchExecuteExtraFormValue, defaultPluginBatchExecutorPageInfo} from "@/defaultConstants/PluginBatchExecutor"
+import {
+    defPluginBatchExecuteExtraFormValue,
+    defaultPluginBatchExecutorPageInfo
+} from "@/defaultConstants/PluginBatchExecutor"
+import {defaultBrutePageInfo} from "@/defaultConstants/NewBrute"
+import {defaultScanPortPageInfo} from "@/defaultConstants/NewPortScan"
+import {defaultPocPageInfo} from "@/defaultConstants/YakPoC"
+import {defaultSpaceEnginePageInfo} from "@/defaultConstants/SpaceEnginePage"
+import { defaultSimpleDetectPageInfo } from "@/defaultConstants/SimpleDetect"
 
 const TabRenameModalContent = React.lazy(() => import("./TabRenameModalContent"))
 const PageItem = React.lazy(() => import("./renderSubPage/RenderSubPage"))
