+++ app/renderer/src/main/src/store/pageInfo.ts
@@ -59,9 +59,9 @@ interface PageParamsInfoProps {
     /**简易版 安全检测页面 */
     simpleDetectPageInfo?: SimpleDetectPageInfoProps
 }
-export const defaultSpaceEnginePageInfo: SpaceEnginePageInfoProps = {}
+
 export interface SpaceEnginePageInfoProps {}
-export const defaultSimpleDetectPageInfo: SimpleDetectPageInfoProps = {}
+
 export interface SimpleDetectPageInfoProps {}
 export interface PluginBatchExecutorPageInfoProps {
     /**执行批量执行的runtimeId */
@@ -93,14 +93,6 @@ export interface WebFuzzerPageInfoProps {
     variableActiveKeys?: string[]
 }
 
-export const defaultPocPageInfo: PocPageInfoProps = {
-    selectGroup: [],
-    selectGroupListByKeyWord: [],
-    formValue: {},
-    https: false,
-    httpFlowIds: [],
-    request: new Uint8Array()
-}
 export interface PocPageInfoProps {
     /** type 1会打开漏洞检测类型选择  2直接带着数据打开poc页面*/
     type?: number
@@ -116,16 +108,12 @@ export interface PocPageInfoProps {
     /**请求包 */
     request: Uint8Array
 }
-export const defaultBrutePageInfo: BrutePageInfoProps = {
-    targets: ""
-}
+
 export interface BrutePageInfoProps {
     /**输入目标 */
     targets: string
 }
-export const defaultScanPortPageInfo: ScanPortPageInfoProps = {
-    targets: ""
-}
+
 export interface ScanPortPageInfoProps {
     /**输入目标 */
     targets: string
